Even backtracking must stop
If an expression backtracked all the way back to index 0
(matching the empty string from 0 to 0), it can no longer
backtrack and match again
For instance, expression `a?a` matched again the empty string as target
Right from the start a? matches the empty string (the whole target)
expression `a` ask a? to backtrack
If a? backtracked (again matching the empty string) then `a` would
endlessly fail and request a? to backtrack
Thus a? must have a `backtrack limit`